Force build to provision its own pip in nox test session (#12848)
Modern versions of build use the pip from the environment it's installed
in, taking advantage of --python to avoid needing to provision pip in
the temporary build environment. This is a neat optimization, but it
breaks when a broken pip is installed from a previous test run.

When this happens, you have to clear out the old Nox session environment
to get build functioning again. This is annoying and clearly works
against the goal of run_with_protected_pip() which is meant to prevent
such issues.

We can force build to provision its own pip by simply uninstalling pip
before invoking the sdist build.