📝💄 Split the "process" change log entry to 3
Previously, the change log had a section called "Process". And some of
the incoming pull requests would add change notes there when they
couldn't be put into other categories. Using it like that is not a
good idea. This patch rethinks the categorization and the audiences of
a few common change types and defines respective categories that
replace "Process".

The new categories are:

  * `packaging` -- news for downstream re-packagers

  * `contrib` -- news for regular project participants

  * `misc` -- public change log for things that don't fit anywhere but
    are useful

Resolves #12555