Rewrite urllib3 retry warnings to be understandable
The approach taken here is to install a logging filter on the _vendor
logging handler which "sniffs" for the appropriate warnings and rewrites
them. This rewriting is best effort because urllib3 gives us scant context
to work with. Under verbose mode, the original error is appended as a
"here's all of the context, sorry that it's unreadable" aid.

This is supported by an upstream PR to urllib3 adding a special extra
attribute to the retrying warning, providing for a clear identifying
marker and extra information needed by pip (i.e. the target host).

Overall, this is quite hacky, but having discussed with the urllib3
maintainers, this is the most mutually acceptable compromise for improving
these user-hostile warnings for pip.