Rewrite urllib3 retry warnings to be understandable, if possible
Alright, so this is 100% a hack, but we don't really have any
alternatives for improving the truly user-hostile warnings that urllib3
emits when retrying after a connection error. Upstream improvements
aren't possible until we are able to catch up with urllib3 2.x, but
we're stuck on the 1.x series until at least 2025 due to OpenSSL
incompatibilties.

Thus, the approach taken here is to install a logging filter on the
connectionpool.py urllib3 module (don't worry, the hacks continue[^1])
which "sniffs" for the appropriate warnings and rewrites them as best as
possible with the limited context urllib3 gives us. Under verbose mode,
the original error is appended as a "here's all of the context, sorry
that it's unreadable" aid.

[^1]: I did have to resort to parsing the timeout error strings with
  regex in order to get the configured timeout (since logically they're
  set per request, not globally or per session... doing it this way
  results in less buggy behaviour and tests a lot less annoying to write).