Implement Dependency Group option: `--group`
`--group` is supported on `download` and `install` commands.
The option is parsed into the more verbose and explicit
`dependency_groups` name on the parsed args.

Both of these commands invoke the same processor for resolving
dependency groups, which loads `pyproject.toml` and resolves the list
of provided groups against the `[dependency-groups]` table.

A small alteration is made to `pip wheel` to initialize
`dependency_groups = []`, as this allows for some lower-level
consistency in the handling of the commands.