Add unit tests for dependency group loading
A new unit test module is added for parsing dependency groups and used
to verify all of the pip-defined behaviors for handling
dependency-groups.

In one path, the underlying exception message from `dependency-groups`
is exposed to users, where it should offer some explanation of why
parsing failed, and this is therefore tested.

Some related changes are applied to the dependency groups usage sites
in the src tree. The signature of the dependency group requirement
parse function is simplified, and its usage is therefore updated.
A bugfix is applied to add a missing `f` on an intended f-string.