Inherit HTTP cache file read/write permissions from cache directory (#13070)
The NamedTemporaryFile class used to create HTTP cache files is hard-coded to use
file mode 600 (owner read/write only). This makes it impossible to share a pip
cache with other users.

With this patch, once a cache file is committed, its permissions are updated to
inherit the read/write permissions of the cache directory. As before, the owner
read/write permissions will always be set to avoid a completely unusable cache.

---------

Co-authored-by: Richard Si <sichard26@gmail.com>