Update `--group` option to take `[path:]name`
In discussions about the correct interface for `pip` to use
[dependency-groups], no strong consensus arose. However, the option
with the most support appears to be to make it possible to pass a file
path plus a group name.

This change converts the `--group` option to take colon-separated
path:groupname pairs, with the path part optional. The CLI parsing
code is responsible for handling the syntax and for filling in a
default path of `"pyproject.toml"`.

If a path is provided, it must have a basename of `pyproject.toml`.
Failing to meet this constraint is an error at arg parsing time.

The `dependency_groups` usage is updated to create a
DependencyGroupResolver per `pyproject.toml` file provided. This
ensures that we only parse each file once, and we keep the results of
previous resolutions when resolving multiple dependency groups from
the same file. (Technically, the implementation is a resolver per
path, which is subtly different from per-file, in that it doesn't
account for symlinks, hardlinks, etc.)