Stop installing pip twice in docs nox sessions
At some point, session.install("pip") in the docs and docs-live nox
sessions was changed to install pip in editable mode, presumably to
enable reruns w/o dependency installation (-R flag) to pick up changes
for our pip sphinx extension. This doesn't do anything though as pip
is reinstalled normally as it's declared in docs/requirements.txt.

I think it's a fair compromise that if you want to pick up changes in
pip's source that show up in the documentation, you should not be using
the -R nox flag.