Pass --proxy to build subprocesses
Similar to --cert and --client-cert, the --proxy flag was not passed
down to the isolated build environment. This was simply an oversight.

I opted to store the original proxy string in a new attribute on the
session as digging into the .proxies dictionary felt janky, and so did
passing the proxy string to the finder as an argument.

Co-authored-by: lcmartin <luis.martinez@collins.com>