Add Dependency Groups to user guide doc
The new section comes after `requirements.txt` and `constraints.txt`
documentation but before documentation about wheels.

The doc attempts to be beginner-friendly and balance
- clarity about the path behavior
- explanation of `[dependency-groups]` itself
- justification of the path-oriented design -- in the form of an
  example of installing from two different subprojects simultaneously

There is therefore ample material not covered in the new section --
e.g., there is no mention of `include-group`, which is explained at
length in the specification doc.