Don't pass --cert to build subprocesses unless also given on CLI
This fixes a regression introduced by commit 34fc0e20bd91f7facd1. After
that patch, --cert would always be given to the nested pip call, either
pointing to pip's CA bundle, or to whatever the user had set on the CLI.
This means truststore is always disabled... which is bad.

We used to have to do some shenanigans to pass the CA bundle to the
subprocess as certifi doesn't (didn't?) really play nice when in a
zipfile. Regardless, we stopped packing pip into a zipfile to provision
the build environment a while ago, so we can simply do the normal thing
and pass --cert when it's actually given. Otherwise, the subprocess will
find its CA bundle without fuss.

There apparently aren't any truststore tests (as testing system CAs is
probably a pain), so I didn't add one here either. At some point, we
should, though.