perf: pass --no-compile for build deps (#13192)
As the build environment is ephemeral, it's wasteful to pre-compile
everything, especially as not every Python module will be used/compiled
in most cases. In addition, the argument is that compilation ensures
read-only installations are not slower than necessary is moot as the
build environment is by design writable.

This saves ~200ms while installing setuptools while building pip itself.
For context, a pip install . (with --no-index and --find-links) takes
3400ms on main and pip wheel . takes 2200ms on main.

So, it's not a major win, but it's also nothing to sneeze at. Packages
using smaller backends, like flit, will likely see smaller improvements.