Add get_console() to query global rich consoles
For an install progress bar, we'd like to emit logs while the progress
bar updates (for uninstallation messages, etc.). To avoid interwoven
logs, we need to log to the same console that the progress bar is using.

This is easiest to achieve by simply storing a global stdout and stderr
console, queried via a get_console() helper.