remove should_build_for_wheel_command
In pip wheel, we always want wheels, of course.
should_build_for_wheel_command was only called when is_wheel is False,
so the logging statement part of _should_build was never reached.
Since this is the only side effect and _should_build otherwise always return
true in these circumstances, it can be removed.