Implement PEP 753 normalization for Home-Page fallback
We treat "homepage" as a well-known project URL, using it to populate
the Home-Page core metadata field if it's undefined. This patch updates
the normalization logic to follow PEP 753.

The logic emitting the raw project URLs was left unchanged as we don't
really care for well-known vs. boring URLs here.