Rewrite flaky test_vcs_url_urlquote_normalization to use local git repo
Looking at the VCS code, I'm not actually sure if the original test is
testing what's supposed to be testing nowadays. The docstring suggests
it's testing the compare_urls() method in vcs/versioncontrol.py but it
never called in the original test.

Anyway, this test has been particularly flaky as Launchpad frequently
timeouts. We don't need to test bzr here, so let's switch to test
against a local git repository.