Provide more information on `import setuptools` failure
In the legacy `setup.py` wrapper importer, failures to import
`setuptools` might not be because `setuptools` itself is absent, but
because it failed to import successfully.  This can currently arise (as
of `setuptools==77.0.1`) in seemingly valid environments due to
limitations in the `setuptools` vendoring.  This commit exposes more of
the failure up to users, to help with debugging failed environments, and
makes it possible for `setuptools` to expose a better error message
itself through the import system.

For some more context, see pypa/setuptools#4894.