Provide traceback on setuptools import failure in legacy paths (#13291)
In the legacy `setup.py` wrapper importer, failures to import
setuptools might not be because setuptools itself is absent, but
because it failed to import successfully.  This can currently arise (as
of `setuptools==77.0.1`) in seemingly valid environments due to
limitations in the setuptools vendoring.  This commit exposes more of
the failure up to users, to help with debugging failed environments, and
makes it possible for setuptools to expose a better error message
itself through the import system.

The above situation is more likely than setuptools being wholly
absent as the legacy code paths invoke setup.py will check that
setuptools is installed (and install setuptools in a temporary build
env. if not) beforehand.

For some more context, see pypa/setuptools#4894.