ci: Test packaging flow on Windows (#13184)
Most of pip's maintainers use non-Windows environments, so it's easy for
our tooling to regress on Windows. Let's avoid that for our release
flow.

In addition, remove the dependency on the packaging job by the test
jobs. The packaging job rarely fails so we aren't saving any CI
resources by failing early if the packaging job fails.