Reverse --no-python-version-warning deprecation (#13308)
This flag is way more prevalent than initially thought. Since this flag
is already a no-op, it's better to hide it from CLI help. This way, we can
avoid unnecessary churn, but avoid/discourage new occurrences of the flag.

The deprecation has already achieved the goal of communicating that
this flag is unnecessary today.