Make double uninstall test independent of order
Discovery order depends on the metadata backend.

For the importlib backend, it has changed following the previous commit,
because now the .egg-link file is found while inspecting site-package,
and the .egg directory is found after because it comes after in sys.path.

Before the dedictated .egg detection logic was triggered while inspecting
site package and found it, before looking for .egg-link.