Introduce resumable downloads with --resume-retries (#12991)
Add the option --resume-retries to allow pip to retry a download X
number of times.

When a download times out or fails midway through and the download
size is known, a HTTP range request will be issued to resume the
download. If the server supports a partial download (i.e. returns 206
Partial Content), the download will be resumed. If the server responds
with 200 (i.e., the server does NOT support range requests or the file
has changed since the last request), we fall back to restarting the
download. This repeats as necessary until the download is complete or
no more retries remain.

The number of resumes currently defaults to zero, but it is expected to
be increased sometime after the 25.1 release once we've gotten feedback.

Signed-off-by: gmargaritis <gmargaritis@protonmail.com>
Co-authored-by: Yichi Yang <helium6072@gmail.com>
Co-authored-by: Richard Si <sichard26@gmail.com>