Trim distlib to remove non scripts code
Distlib is our largest dependency. The vast majority of that is the
Windows stub executables which we can't get rid of. However, distlib
is a general purpose low-level packaging library, of which we only
use the scripts module.

By line count, trimming the parts of distlib that aren't used for
scripts results in a 5% reduction in total (`_internal` and `_vendor`)
size, which translates to ~250 KB of uncompressed and 64 KB of
compressed (wheel) savings.