Remove LRU cache from methods [ruff rule cached-instance-method] (#13306)
Enables ruff rule cached-instance-method (B019).

By caching instance methods, the instances are kept alive until they are
evicted out of the LRU cache. If the cached instance is itself is
referencing other memory-heavy objects, this can lead to hard-to-debug
high memory usage.

Most of the objects fixed here are fairly long lived, but we don't
necessarily want them to be kept alive until the very end of the process.
Hypothetically, if the build isolation became in-process there will be
multiple instances of these classes that should only live for the
lifetime of any particular build isolation.