feat(vcs/git): add verbosity argument for Git submodule command
The verbosity argument enables callers to control the supression of command
output emitted by the Git submodule command. This is helpful for situations
where the Git command call may take a long time, such as with many submodules
registered in a repository, or a slow internet connection.

The `verbosity` argument has, different than some of its caller methods, a
boolean type for verbosity, as the options for verbosity are binary.  Verbosity
can only be supressed via the `--quiet` flag of the `git submodule update`
command.

Implements: https://github.com/pypa/pip/issues/13329