feat(vcs): add verbosity argument to update
verbosity argument to the update method allows for controlling the supression of
output executed from this method. The argument should be a
boolean value as the `-q` flag is binary, but it now is of type int as
to not require a refactor of the base class callers.

Implements: https://github.com/pypa/pip/issues/13329