feat(vcs): add verbosity argument to switching
verbosity argument to the update method allows for controlling the supression of
output by commands executed from this method. The argument should be a
boolean value as the `-q` flag is binary, but it now is of type int as
to not require a major refactor of the base class, should other inherited
classes also implement the extension of verbosity.

Implements: https://github.com/pypa/pip/issues/13329