pylock: preserve distinction between absent and empty extras and depdency-groups
The specs says these fields are optional with default = [].
But it also says "Tools SHOULD explicitly set this key to an empty array to signal that the inputs used to generate the lock file had no dependency groups (e.g. a pyproject.toml file had no [dependency-groups] table), signalling that the lock file is, in effect, multi-use even if it only looks to be single-use.", implying that
the distinction between absent and empty is meaningful.