Fix if-range handling and its unit tests
The value used for conditional range requests should be updated when the
download is restarted (not *resumed*). The unit tests testing
conditional range requests were broken, using code 206 for the
non-matching responses. This violates the HTTP spec and fools the
resumption logic into thinking the conditional range request was
fulfilled (when it wasn't!).

The test cases were fixed to use code 200 as expected and extended
to ensure the content key is updated across a chain of non-matching
repsonses (which is borderline impossible, but hey it did catch my
eye so why not fix it).