Handle pathologically connected dependency graphs
When trying to install a set of packages which have a pathological
dependency graph, potentially with lots of cycles, walking the graph
exhaustively results in exponential runtime. The aim of that walk was to
determine which packages are the most "leaf-like" ones, being furthest
away from the root package. Instead we use a heuristic to determine a
good way to break the cycle in the graph.