Limit the DFS for pathologically connected graphs
Since the DFS of the dependency graph is brute force it is exponential in the
number of children. If a graph is pathological in nature (i.e. many
cross-dependencies) then this will not complete in "reasonable" time. To address
this we limit the number of visits to any node in the graph, thus capping the
complexity to be linear in nature.