Re-enable blank issues
The TL;DR is that GitHub is enforcing blank_issues_enabled=false in the issue template configuration strictly (we used to use a direct URL as a bypass). This is a recent change. The core team uses blank issues extensively so we need to reenable blank issues. Unfortunately, this has the side effect of allowing everyone to file blank issues. We'll simply have to see how much (if any) misfiled issues we get.