docs nox session: delete old build beforehand (#13471)
While Sphinx is good at only rebuilding what was changed, it often
doesn't rebuild *enough*. For example, if you update a document that
is referenced by another, the embedded reference title in the latter
is not updated on the next build. This also can hide warnings (which
then fail the docs CI ... which is quite annoying).

To fix this, simply delete the old docs build output entirely
beforehand. If a developer wants a fast edit-build-view cycle, then the
docs-live session is what they should be using.