test: Enforce network hygiene with pytest-subket (#13481)
pytest-subket is my fork of pytest-socket with changes that allow it to
function in subprocesses and even across environments (as long as the
pytest-subket package is installed, no dependencies needed).

To promote better test hygiene, tests should need to opt-in to access
the Internet. This is managed via the network marker, but is not
enforced. With pytest-subket, tests accessing the Internet w/o the
marker will immediately fail.

This will encourage us to write tests to avoid hitting the network which
will result in a faster, more reliable test suite that won't regress.
This also makes the lives of downstream that run our test suite in a
sandboxed environment nicer as our network marker should be up to date
24/7 now (barring any tests using remote VCS repositories).

* test: Bring network marker up to date with reality

All instances of the network marker were removed. Afterwards a full test
suite was performed with pytest-subket active. The marker was reapplied
to the tests that actually needed it (including tests that hit the
network via Git or other VCS software).