Support Direct URL editable requirements
This is crucial for the deprecation of non-bare egg URL fragments as they
used to be the sole way to request an extra for a VCS editable install.

I've also removed all references to egg fragments from the user docs and
added further explanation of the Direct URL form in the VCS support
topic.

Getting the error handling right was a pain. That's why I extracted the
two parsing flows into their own functions.

Co-authored-by: Tzu-ping Chung <uranusjr@gmail.com>