Fix broken unit tests & xfail weird failure
Fixed the test_unsupported_hashes and test_install_editable_with_wrong_egg_name
tests to use pip's own path -> URL function instead of simply adding file://
which results in the malformed file://C:/. This will be handled gracefully in
the next 3.14 release (as a matter of backwards compatibility) but the test
suite shouldn't be relying on such behaviour anyway.

Marked test_build_deps_use_proxy_from_cli as a xfail on Python 3.14+
because it seems to be an intermittent issue with proxy.py or something
else with our test suite. I can only reproduce the failure locally if I
run the test suite with two workers 🙃. Given it doesn't seem to be a
problem with pip itself, I'm prioritizing getting 25.2 released on time
than doing a full proper investigation into what's wrong.