Ensure selfcheck file inherits directory permissions (#13528)
When writing files to the cache, pip copies the permissions from the cache
directory. This is an expected behaviour to ensure that any user that has
access to the cache has also access to its content.

Copying the permissions explicitly is necessary due to the use of
adjacent_tmp_file() which by virtue of tempfile.mkstemp() (indirectly via
NamedTemporaryFile) causes all files to be created with 600
permissions.

This is not happening for the selfcheck cache, rendering it unaccessible
by users different from the one that first created the cache.

Signed-off-by: Alessandro Molina <alessandro@molina.fyi>
Co-authored-by: Richard Si <sichard26@gmail.com>