Require the revision to be percent-encoded in VSC URLs (#13407)
This changes how the revision part is handled in version control URLs.
When a branch name contains special characters such as '@' or '#', they
must be percent-encoded in URLs. This is backwards incompatible, but
the fallout should be negligible. In most cases, it will result in an
one-time error. It's extremely unlikely that people have percent-encoding
in their branch names (which could be misinterpreted).

It is also extremely unlikely that someone has both branches with and
without percent-escaping names that could cause an unnoticed behavioural
change.