Implement fallthrough logic for options
With the merging of PR #12201, an additional layering was added to the
variants in the Configuration, namely to hold the filename from which
this variant was read.

This extra layer broke the original fallthrough logic of being able to
overrule configuration options in the global config with back to default
settings from the user config.

This PR re-implements the fallthrough logic, but now in the
_get_ordered_configuration_items in the ConfigOptionParser. Instead of
constructing the section_items directly as a list of key-value tuples
per section, we first create a dictionary of key-value items per section
(which ensures we just keep one value per key) and then convert this in
one go to the original dictionary per section of a key-value list