test: Pre-compile common wheels in base test env
The test suite disables Python bytecode caching so if we don't
pre-compile the bytecode for coverage, setuptools, and pytest-subket
which are all preinstalled in the test environments, we pay a heavy
import time penalty (e.g., setuptools during a PEP 517 build). pip is
already pre-compiled, FWIW.

This is doubly bad for coverage and pytest-subket since they are
imported on every Python process startup.

Locally, this shaves a full minute from a full test suite run down to
2:45 from 3:45.