Fix tests on systems where platlib and purelib are different (#13727)
The following tests expect installed files to be in purelib:

* `tests/functional/test_check.py::test_check_unsupported`
* `tests/functional/test_install_wheel.py::test_basic_install_from_wheel`

However, the actual wheels they install have `Root-Is-Purelib` set to
`false` (or missing entirely), which causes the contents to be installed to
platlib instead. Since the wheels in question don't have any
platform-dependent files, it seems appropriate to update them to use
`Root-Is-Purelib: true`.

Since we're touching `has.script-1.0-py2.py3-none-any.whl`, fix the `Tag`
field as well, which was set incorrectly (to `cp26-none-linux_x86_64`).